Splitting Documents.
