+++ api_v2/views/__init__.py
@@ -6,3 +6,8 @@ from .creature import CreatureFilterSet, CreatureViewSet
 from .creature import CreatureTypeViewSet
 
 from .background import BackgroundFilterSet, BackgroundViewSet
+
+from .document import DocumentViewSet
+from .document import RulesetViewSet
+from .document import PublisherViewSet
+from .document import LicenseViewSet
